Squashed commit of the following:
commit 1368b26437abd7b2963374643e3e40d57622749f
Author: ptlthg <24925519+ptlthg@users.noreply.github.com>
Date:   Wed Jul 16 00:23:06 2025 -0400

    Fix circular dependencies

commit 9bf1c0a9e59b91a3303d86d6a9b610d7b840f7e0
Author: ptlthg <24925519+ptlthg@users.noreply.github.com>
Date:   Sat Jul 12 02:01:54 2025 -0400

    Fix armor set selection resetting
