+++ src/constants/upgrades.ts
@@ -1,10 +1,10 @@
-import { EliteItemDto } from '../fortune/item.js';
-import { UpgradeableInfo } from '../fortune/upgradeable.js';
+import type { EliteItemDto } from '../fortune/item.js';
+import type { UpgradeableInfo } from '../fortune/upgradeable.js';
 import type { GearSlot } from '../items/armor.js';
 import type { FARMING_TOOLS } from '../items/tools.js';
-import { JacobContestMedal } from '../util/jacob.js';
-import { Crop } from './crops.js';
-import { Stat } from './stats.js';
+import type { JacobContestMedal } from '../util/jacob.js';
+import type { Crop } from './crops.js';
+import type { Stat } from './stats.js';
 
 export enum UpgradeReason {
 	NextTier = 'next', // Standard upgrade
