+++ src/fortune/farmingarmor.ts
@@ -1,35 +1,31 @@
-import { Crop } from '../constants/crops.js';
+import type { Crop } from '../constants/crops.js';
 import { FARMING_ENCHANTS } from '../constants/enchants.js';
-import { Rarity, Reforge, ReforgeTarget, ReforgeTier } from '../constants/reforges.js';
+import { type Rarity, type Reforge, ReforgeTarget, type ReforgeTier } from '../constants/reforges.js';
 import { Skill } from '../constants/skills.js';
-import { MATCHING_SPECIAL_CROP, SpecialCrop } from '../constants/specialcrops.js';
+import { MATCHING_SPECIAL_CROP, type SpecialCrop } from '../constants/specialcrops.js';
 import { Stat } from '../constants/stats.js';
-import { FortuneSourceProgress, FortuneUpgrade } from '../constants/upgrades.js';
+import type { FortuneSourceProgress, FortuneUpgrade } from '../constants/upgrades.js';
 import { calculateAverageSpecialCrops } from '../crops/special.js';
 import {
 	ARMOR_SET_BONUS,
-	ArmorSetBonus,
+	type ArmorSetBonus,
 	FARMING_ARMOR_INFO,
-	FarmingArmorInfo,
+	type FarmingArmorInfo,
 	GEAR_SLOTS,
 	GearSlot,
 } from '../items/armor.js';
-import { FARMING_EQUIPMENT_INFO } from '../items/equipment.js';
-import { PlayerOptions } from '../player/playeroptions.js';
+import type { PlayerOptions } from '../player/playeroptions.js';
+import { getSourceProgress } from '../upgrades/getsourceprogress.js';
+import { getFakeItem, registerItem } from '../upgrades/itemregistry.js';
 import { ARMOR_SET_FORTUNE_SOURCES } from '../upgrades/sources/armorsetsources.js';
 import { GEAR_FORTUNE_SOURCES } from '../upgrades/sources/gearsources.js';
-import {
-	getItemUpgrades,
-	getLastItemUpgradeableTo,
-	getSelfFortuneUpgrade,
-	getSourceProgress,
-	getUpgradeableRarityUpgrade,
-} from '../upgrades/upgrades.js';
+import { getLastItemUpgradeableTo, getSelfFortuneUpgrade, getUpgradeableRarityUpgrade } from '../upgrades/upgrades.js';
 import { getFortuneFromEnchant } from '../util/enchants.js';
 import { getPeridotFortune } from '../util/gems.js';
 import { FarmingEquipment } from './farmingequipment.js';
-import { EliteItemDto } from './item.js';
-import { UpgradeableBase, UpgradeableInfo } from './upgradeable.js';
+import type { EliteItemDto } from './item.js';
+import type { UpgradeableInfo } from './upgradeable.js';
+import { UpgradeableBase } from './upgradeablebase.js';
 
 export interface ActiveArmorSetBonus {
 	count: number;
@@ -120,7 +116,9 @@ export class ArmorSet {
 		for (const piece of this.equipmentPieces) {
 			piece.setOptions(options);
 		}
+	}
 
+	resetChosenPieces() {
 		this.setArmor(this.pieces);
 		this.setEquipment(this.equipmentPieces);
 	}
@@ -320,10 +318,8 @@ export class ArmorSet {
 		if (!piece) {
 			piece =
 				GEAR_SLOTS[slot].target === ReforgeTarget.Armor
-					? FarmingArmor.fakeItem(FARMING_ARMOR_INFO[GEAR_SLOTS[slot].startingItem] as UpgradeableInfo)
-					: FarmingEquipment.fakeItem(
-							FARMING_EQUIPMENT_INFO[GEAR_SLOTS[slot].startingItem] as UpgradeableInfo
-						);
+					? getFakeItem<FarmingArmor>(GEAR_SLOTS[slot].startingItem)
+					: getFakeItem<FarmingEquipment>(GEAR_SLOTS[slot].startingItem);
 			return piece?.getProgress(true) ?? [];
 		}
 
@@ -511,4 +507,12 @@ export class FarmingArmor extends UpgradeableBase {
 
 		return new FarmingArmor(fake, options);
 	}
+}
+
+for (const item of Object.values(FARMING_ARMOR_INFO)) {
+	if (!item) continue;
+	registerItem({
+		info: item,
+		fakeItem: (i, o) => FarmingArmor.fakeItem(i, o),
+	});
 }
