+++ src/fortune/upgradeable.ts
@@ -1,11 +1,9 @@
-import { Crop } from '../constants/crops.js';
-import { REFORGES, Rarity, RarityRecord, Reforge, ReforgeTarget, ReforgeTier } from '../constants/reforges.js';
-import { Stat, StatsRecord } from '../constants/stats.js';
-import { FortuneSourceProgress, FortuneUpgrade, Upgrade, UpgradeCost } from '../constants/upgrades.js';
-import { PlayerOptions } from '../player/playeroptions.js';
-import { getItemUpgrades, getLastItemUpgradeableTo, getNextItemUpgradeableTo } from '../upgrades/upgrades.js';
-import { getRarityFromLore, previousRarity } from '../util/itemstats.js';
-import { EliteItemDto } from './item.js';
+import type { Crop } from '../constants/crops.js';
+import type { Rarity, RarityRecord, Reforge, ReforgeTarget, ReforgeTier } from '../constants/reforges.js';
+import type { Stat, StatsRecord } from '../constants/stats.js';
+import type { FortuneSourceProgress, FortuneUpgrade, Upgrade, UpgradeCost } from '../constants/upgrades.js';
+import type { PlayerOptions } from '../player/playeroptions.js';
+import type { EliteItemDto } from './item.js';
 
 export type GemSlotCost = { type: 'ITEM'; item_id: string; amount: number } | { type: 'COINS'; coins: number };
 
@@ -41,86 +39,4 @@ export interface Upgradeable {
 	getItemUpgrade(): Upgrade | undefined;
 	getLastItemUpgrade(): { upgrade: Upgrade; info: UpgradeableInfo } | undefined;
 	getProgress(zeroed: boolean): FortuneSourceProgress[];
-}
-
-export class UpgradeableBase implements Upgradeable {
-	public item: EliteItemDto;
-	public info: UpgradeableInfo;
-
-	private _type?: ReforgeTarget | undefined;
-	public get type(): ReforgeTarget | undefined {
-		return this._type;
-	}
-	public set type(value: ReforgeTarget | undefined) {
-		this._type = value;
-	}
-
-	public crop?: Crop;
-	public options?: PlayerOptions;
-
-	public recombobulated: boolean;
-	public rarity: Rarity;
-	public reforge?: Reforge | undefined;
-	public reforgeStats?: ReforgeTier | undefined;
-
-	public fortune: number;
-
-	private items: Partial<Record<string, UpgradeableInfo>>;
-
-	constructor(options: {
-		item: EliteItemDto;
-		options?: PlayerOptions;
-		items: Partial<Record<string, UpgradeableInfo>>;
-	}) {
-		this.item = options.item;
-		const info = options.items[options.item.skyblockId as keyof typeof options.items];
-		if (!info) {
-			throw new Error(`Unknown item: ${options.item.name} (${options.item.skyblockId})`);
-		}
-		this.info = info;
-		this.options = options.options;
-		this.items = options.items;
-
-		if (this.item.lore) {
-			this.rarity = getRarityFromLore(this.item.lore);
-		} else {
-			this.rarity = previousRarity(this.info.maxRarity);
-			this.rarity ??= Rarity.Common;
-		}
-
-		this.reforge = REFORGES[options.item.attributes?.modifier ?? ''] ?? undefined;
-		this.reforgeStats = this.reforge?.tiers?.[this.rarity];
-
-		this.recombobulated = this.item.attributes?.rarity_upgrades === '1';
-		this.fortune = 0;
-	}
-
-	getFortune(): number {
-		return this.fortune;
-	}
-
-	getCalculatedStats(): Partial<Record<Stat, number>> {
-		if (!this.info.computedStats || !this.options) return {};
-		return this.info.computedStats?.(this.options) ?? {};
-	}
-
-	getUpgrades(): FortuneUpgrade[] {
-		return getItemUpgrades(this);
-	}
-
-	getProgress(): FortuneSourceProgress[] {
-		return [];
-	}
-
-	getItemUpgrade(): Upgrade | undefined {
-		return this.info.upgrade;
-	}
-
-	getNextItemUpgrade(): { upgrade: Upgrade; info: UpgradeableInfo } | undefined {
-		return getNextItemUpgradeableTo(this, this.items);
-	}
-
-	getLastItemUpgrade(): { upgrade: Upgrade; info: UpgradeableInfo } | undefined {
-		return getLastItemUpgradeableTo(this, this.items);
-	}
 }
