+++ src/items/equipment.ts
@@ -2,7 +2,7 @@ import { Rarity } from '../constants/reforges.js';
 import { Skill } from '../constants/skills.js';
 import { Stat } from '../constants/stats.js';
 import { UpgradeReason } from '../constants/upgrades.js';
-import { FarmingArmorInfo, GearSlot } from './armor.js';
+import { type FarmingArmorInfo, GearSlot } from './armor.js';
 
 export const FARMING_EQUIPMENT_INFO: Partial<Record<string, FarmingArmorInfo>> = {
 	LOTUS_BRACELET: {
