+++ src/upgrades/getfortune.ts
@@ -1,0 +1,5 @@
+import type { FortuneSource } from '../constants/upgrades.js';
+
+export function getFortune(level: number | null | undefined, source: FortuneSource) {
+	return Math.min(Math.max(level ?? 0, 0), source.maxLevel) * source.fortunePerLevel;
+}
