+++ src/upgrades/sources/accessorysources.ts
@@ -1,10 +1,10 @@
 import { Rarity } from '../../constants/reforges.js';
 import { Stat } from '../../constants/stats.js';
-import { FarmingAccessory } from '../../fortune/farmingaccessory.js';
+import type { FarmingAccessory } from '../../fortune/farmingaccessory.js';
 import { GemRarity } from '../../fortune/item.js';
 import { getPeridotFortune, getPeridotGemFortune } from '../../util/gems.js';
 import { getUpgradeableGems } from '../upgrades.js';
-import { DynamicFortuneSource } from './toolsources.js';
+import type { DynamicFortuneSource } from './toolsources.js';
 
 export const ACCESSORY_FORTUNE_SOURCES: DynamicFortuneSource<FarmingAccessory>[] = [
 	{
